optimized persistence

+++ src/main/java/telran/java2022/person/dto/CityPopulationDto.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class CityPopulationDto {
 	String city;
-	int population;
+	long population;
 }
